requirements.txt
wsgiref==0.1.2
gunicorn==19.0.0
Django==1.6.5
django-chunked-upload
django-jwt-auth
djangorestframework-jwt
djangorestframework
